# packages/api/generated/.docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: Vault Hub Server
  version: 1.0.6
servers:
- url: /
paths:
...
  /api/api-keys:
    get:
      description: Get API keys for the current user with pagination
      operationId: getAPIKeys
      parameters:
      - description: "Number of API keys per page (default 20, max 1000)"
        explode: true
        in: query
        name: pageSize
        required: true
        schema:
          default: 20
          maximum: 1000
          minimum: 1
          type: integer
        style: form
      - description: "Page index, starting from 1 (default 1)"
        explode: true
        in: query
        name: pageIndex
        required: true
        schema:
          default: 1
          minimum: 1
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIKeysResponse"
          description: List of API keys
      tags:
      - APIKey
    post:
      description: Create a new API key
      operationId: createAPIKey
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAPIKeyRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateAPIKeyResponse"
          description: API key created successfully
      tags:
      - APIKey
  /api/api-keys/{id}:
    delete:
      description: Delete an API key
      operationId: deleteAPIKey
      parameters:
      - description: API Key ID
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "204":
          description: API key deleted successfully
      tags:
      - APIKey
    patch:
      description: Update an API key (enable/disable or modify properties)
      operationId: updateAPIKey
      parameters:
      - description: API Key ID
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAPIKeyRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VaultAPIKey"
          description: API key updated successfully
      tags:
      - APIKey
  /api/cli/vaults:
    get:
      description: Get all vaults accessible by API key
      operationId: getVaultsByAPIKey
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/VaultLite"
                type: array
          description: List of vaults accessible by API key
      security:
      - ApiKeyAuth: []
      tags:
      - Cli
  /api/cli/vault/{uniqueId}:
    get:
      description: Get a specific vault by Unique ID using API key
      operationId: getVaultByAPIKey
      parameters:
      - description: Vault Unique ID
        explode: false
        in: path
        name: uniqueId
        required: true
        schema:
          type: string
        style: simple
      - description: Enable client-side encryption (server returns encrypted value)
        explode: false
        in: header
        name: X-Enable-Client-Encryption
        required: false
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vault"
          description: Vault details
      security:
      - ApiKeyAuth: []
      tags:
      - Cli
  /api/cli/vault/name/{name}:
    get:
      description: Get a specific vault by name using API key
      operationId: getVaultByNameAPIKey
      parameters:
      - description: Vault name
        explode: false
        in: path
        name: name
        required: true
        schema:
          type: string
        style: simple
      - description: Enable client-side encryption (server returns encrypted value)
        explode: false
        in: header
        name: X-Enable-Client-Encryption
        required: false
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vault"
          description: Vault details
      security:
      - ApiKeyAuth: []
      tags:
      - Cli
components:
  schemas:
...
    VaultLite:
      example:
        name: name
        description: description
        category: category
        uniqueId: uniqueId
        updatedAt: 2000-01-23T04:56:07.000+00:00
      properties:
        uniqueId:
          description: Unique identifier for the vault
          type: string
        name:
          description: Human-readable name
          type: string
        description:
          description: Human-readable description
          type: string
        category:
          description: Category/type of vault
          type: string
        updatedAt:
          format: date-time
          type: string
      required:
      - name
      - uniqueId
      type: object
    Vault:
      example:
        createdAt: 2000-01-23T04:56:07.000+00:00
        name: name
        description: description
        category: category
        userId: 0
        value: value
        uniqueId: uniqueId
        updatedAt: 2000-01-23T04:56:07.000+00:00
      properties:
        uniqueId:
          description: Unique identifier for the vault
          type: string
        userId:
          description: ID of the user who owns this vault
          format: int64
          type: integer
        name:
          description: Human-readable name
          type: string
        value:
          description: Encrypted value
          type: string
        description:
          description: Human-readable description
          type: string
        category:
          description: Category/type of vault
          type: string
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
      required:
      - name
      - uniqueId
      - value
      type: object
...
    APIKeysResponse:
      example:
        pageIndex: 1
        apiKeys:
        - createdAt: 2000-01-23T04:56:07.000+00:00
          lastUsedAt: 2000-01-23T04:56:07.000+00:00
          name: name
          vaults:
          - name: name
            description: description
            category: category
            uniqueId: uniqueId
            updatedAt: 2000-01-23T04:56:07.000+00:00
          - name: name
            description: description
            category: category
            uniqueId: uniqueId
            updatedAt: 2000-01-23T04:56:07.000+00:00
          id: 0
          isActive: true
          expiresAt: 2000-01-23T04:56:07.000+00:00
          updatedAt: 2000-01-23T04:56:07.000+00:00
        - createdAt: 2000-01-23T04:56:07.000+00:00
          lastUsedAt: 2000-01-23T04:56:07.000+00:00
          name: name
          vaults:
          - name: name
            description: description
            category: category
            uniqueId: uniqueId
            updatedAt: 2000-01-23T04:56:07.000+00:00
          - name: name
            description: description
            category: category
            uniqueId: uniqueId
            updatedAt: 2000-01-23T04:56:07.000+00:00
          id: 0
          isActive: true
          expiresAt: 2000-01-23T04:56:07.000+00:00
          updatedAt: 2000-01-23T04:56:07.000+00:00
        pageSize: 6
        totalCount: 0
      properties:
        apiKeys:
          items:
            $ref: "#/components/schemas/VaultAPIKey"
          type: array
        totalCount:
          description: Total number of API keys
          type: integer
        pageSize:
          description: Number of API keys per page
          type: integer
        pageIndex:
          description: Current page index (starting from 1)
          type: integer
      required:
      - apiKeys
      - pageIndex
      - pageSize
      - totalCount
      type: object
    VaultAPIKey:
      example:
        createdAt: 2000-01-23T04:56:07.000+00:00
        lastUsedAt: 2000-01-23T04:56:07.000+00:00
        name: name
        vaults:
        - name: name
          description: description
          category: category
          uniqueId: uniqueId
          updatedAt: 2000-01-23T04:56:07.000+00:00
        - name: name
          description: description
          category: category
          uniqueId: uniqueId
          updatedAt: 2000-01-23T04:56:07.000+00:00
        id: 0
        isActive: true
        expiresAt: 2000-01-23T04:56:07.000+00:00
        updatedAt: 2000-01-23T04:56:07.000+00:00
      properties:
        id:
          description: Unique API key ID
          format: int64
          type: integer
        name:
          description: Human-readable name for the API key
          type: string
        vaults:
          description: Array of vaults this key can access (null/empty = all user's
            vaults)
          items:
            $ref: "#/components/schemas/VaultLite"
          type: array
        expiresAt:
          description: Optional expiration date
          format: date-time
          type: string
        lastUsedAt:
          description: When the key was last used
          format: date-time
          type: string
        isActive:
          description: Whether the key is currently active
          type: boolean
        createdAt:
          description: When the key was created
          format: date-time
          type: string
        updatedAt:
          description: When the key was last updated
          format: date-time
          type: string
      required:
      - createdAt
      - id
      - isActive
      - name
      type: object
    CreateAPIKeyRequest:
      properties:
        name:
          description: Human-readable name for the API key
          maxLength: 255
          minLength: 1
          type: string
        vaultUniqueIds:
          description: Array of vault unique IDs this key can access (empty = all
            user's vaults)
          items:
            type: string
          type: array
        expiresAt:
          description: Optional expiration date
          format: date-time
          type: string
      required:
      - name
      type: object
    CreateAPIKeyResponse:
      example:
        apiKey:
          createdAt: 2000-01-23T04:56:07.000+00:00
          lastUsedAt: 2000-01-23T04:56:07.000+00:00
          name: name
          vaults:
          - name: name
            description: description
            category: category
            uniqueId: uniqueId
            updatedAt: 2000-01-23T04:56:07.000+00:00
          - name: name
            description: description
            category: category
            uniqueId: uniqueId
            updatedAt: 2000-01-23T04:56:07.000+00:00
          id: 0
          isActive: true
          expiresAt: 2000-01-23T04:56:07.000+00:00
          updatedAt: 2000-01-23T04:56:07.000+00:00
        key: key
      properties:
        apiKey:
          $ref: "#/components/schemas/VaultAPIKey"
        key:
          description: The generated API key (only shown once)
          type: string
      required:
      - apiKey
      - key
      type: object
    UpdateAPIKeyRequest:
      properties:
        name:
          description: Human-readable name for the API key
          maxLength: 255
          minLength: 1
          type: string
        vaultUniqueIds:
          description: Array of vault unique IDs this key can access (empty = all
            user's vaults)
          items:
            type: string
          type: array
        expiresAt:
          description: Optional expiration date
          format: date-time
          type: string
      type: object
...
  securitySchemes:
    apiKeyAuth:
      scheme: bearer
      type: http