# packages/api/generated/handlers/api_cli.go
package handlers
import (
	"github.com/GIT_USER_ID/GIT_REPO_ID/models"
	"github.com/labstack/echo/v4"
	"net/http"
)

// GetVaultByAPIKey - 
func (c *Container) GetVaultByAPIKey(ctx echo.Context) error {
	return ctx.JSON(http.StatusOK, models.HelloWorld {
		Message: "Hello World",
	})
}


// GetVaultByNameAPIKey - 
func (c *Container) GetVaultByNameAPIKey(ctx echo.Context) error {
	return ctx.JSON(http.StatusOK, models.HelloWorld {
		Message: "Hello World",
	})
}


// GetVaultsByAPIKey - 
func (c *Container) GetVaultsByAPIKey(ctx echo.Context) error {
	return ctx.JSON(http.StatusOK, models.HelloWorld {
		Message: "Hello World",
	})
}