# packages/api/generated_handlers/api_api_key.go
package generated_handlers

import (
	"github.com/labstack/echo/v4"
	"github.com/lwshen/vault-hub/packages/api/generated_models"
	"net/http"
)

// CreateAPIKey -
func (c *Container) CreateAPIKey(ctx echo.Context) error {
	return ctx.JSON(http.StatusOK, generated_models.HelloWorld{
		Message: "Hello World",
	})
}

// DeleteAPIKey -
func (c *Container) DeleteAPIKey(ctx echo.Context) error {
	return ctx.JSON(http.StatusOK, generated_models.HelloWorld{
		Message: "Hello World",
	})
}

// GetAPIKeys -
func (c *Container) GetAPIKeys(ctx echo.Context) error {
	return ctx.JSON(http.StatusOK, generated_models.HelloWorld{
		Message: "Hello World",
	})
}

// UpdateAPIKey -
func (c *Container) UpdateAPIKey(ctx echo.Context) error {
	return ctx.JSON(http.StatusOK, generated_models.HelloWorld{
		Message: "Hello World",
	})
}